Centralise compiling stuff and requirements
