+++ build-all.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+./build-giants.py && audioquake/build-audioquake.py
