+++ ldl/.gitignore
@@ -1,11 +1,7 @@
 *.o
-qbsp
-light
-vis
 quake.wad
 *.bsp
 ldl-tutorial.html
-bin/
 
 # Built map files
 first_20080427_1353.map
