--- ldl/requirements.txt
@@ -1,3 +1,0 @@
-argcomplete==1.10.0
-xmltodict==0.12.0
-patch_ng==1.17
