+++ requirements.txt
@@ -4,3 +4,7 @@ pywin32==228; sys_platform=="win32"
 mistune==0.8.4
 mistune-contrib==0.1
 PyInstaller==3.6
+
+argcomplete==1.10.0
+xmltodict==0.12.0
+patch_ng==1.17
